add pr feedback and remove unused files
