+++ .github/workflows/ci.yaml
@@ -96,7 +96,7 @@ jobs:
       fail-fast: false
       matrix:
         tox-environments:
-          - integration-ha-replication
+          - integration-ha
     name: ${{ matrix.tox-environments }}
     needs:
       - lint
