+++ tests/integration/ha/test_ha.py
@@ -4,14 +4,17 @@
 
 import asyncio
 import logging
+import time
 
 import pytest
-from helpers import (
+from ha_helpers import (
     APP_NAME,
     REL_NAME_ADMIN,
     ZK_NAME,
     check_logs,
+    get_topic_leader,
     produce_and_check_logs,
+    send_control_signal,
 )
 from pytest_operator.plugin import OpsTest
 
@@ -22,12 +25,17 @@ DUMMY_NAME = "app"
 
 
 @pytest.mark.abort_on_fail
-async def test_build_and_deploy(ops_test: OpsTest, kafka_charm):
+async def test_build_and_deploy(ops_test: OpsTest, kafka_charm, app_charm):
     await asyncio.gather(
         ops_test.model.deploy(kafka_charm, application_name=APP_NAME, num_units=1, series="jammy"),
         ops_test.model.deploy(ZK_NAME, channel="edge", num_units=1, series="jammy"),
+        ops_test.model.deploy(app_charm, application_name=DUMMY_NAME, num_units=1, series="jammy"),
     )
-    await ops_test.model.wait_for_idle(apps=[APP_NAME, ZK_NAME], idle_period=30, timeout=3600)
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, ZK_NAME, DUMMY_NAME],
+        idle_period=30,
+        timeout=3600,
+    )
     assert ops_test.model.applications[APP_NAME].status == "blocked"
     assert ops_test.model.applications[ZK_NAME].status == "active"
 
@@ -37,8 +45,61 @@ async def test_build_and_deploy(ops_test: OpsTest, kafka_charm):
         assert ops_test.model.applications[APP_NAME].status == "active"
         assert ops_test.model.applications[ZK_NAME].status == "active"
 
+    await ops_test.model.add_relation(APP_NAME, f"{DUMMY_NAME}:{REL_NAME_ADMIN}")
+    await ops_test.model.wait_for_idle(apps=[APP_NAME, DUMMY_NAME], idle_period=30)
+    assert ops_test.model.applications[APP_NAME].status == "active"
+    assert ops_test.model.applications[DUMMY_NAME].status == "active"
 
-async def test_second_cluster(ops_test: OpsTest, kafka_charm, app_charm):
+
+async def test_replicated_events(ops_test: OpsTest):
+    await ops_test.model.applications[APP_NAME].add_units(count=2)
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME], status="active", timeout=600, idle_period=120, wait_for_exact_units=3
+    )
+    logger.info("Producing messages and checking on all units")
+    produce_and_check_logs(
+        model_full_name=ops_test.model_full_name,
+        kafka_unit_name=f"{APP_NAME}/0",
+        provider_unit_name=f"{DUMMY_NAME}/0",
+        topic="replicated-topic",
+        replication_factor=3,
+        num_partitions=1,
+    )
+    # check logs in the two remaining units
+    check_logs(
+        model_full_name=ops_test.model_full_name,
+        kafka_unit_name=f"{APP_NAME}/1",
+        topic="replicated-topic",
+    )
+    check_logs(
+        model_full_name=ops_test.model_full_name,
+        kafka_unit_name=f"{APP_NAME}/2",
+        topic="replicated-topic",
+    )
+
+
+async def test_kill_broker_with_topic_leader(ops_test: OpsTest):
+    initial_leader_num = await get_topic_leader(ops_test=ops_test, topic="replicated-topic")
+    logger.info(f"Killing broker of leader for topic 'replicated-topic': {leader_num}")
+    await send_control_signal(
+        ops_test=ops_test, unit_name=f"{APP_NAME}/{leader_num}", kill_code="SIGKILL"
+    )
+    # Give time for the service to restart
+    time.sleep(15)
+    # Check that is still possible to write to the same topic.
+    final_leader_num = await get_topic_leader(ops_test=ops_test, topic="replicated-topic")
+    assert initial_leader_num != final_leader_num
+
+    produce_and_check_logs(
+        model_full_name=ops_test.model_full_name,
+        kafka_unit_name=f"{APP_NAME}/0",
+        provider_unit_name=f"{DUMMY_NAME}/0",
+        topic="replicated-topic",
+        create_topic=False,
+    )
+
+
+async def test_multi_cluster_isolation(ops_test: OpsTest, kafka_charm):
     second_kafka_name = f"{APP_NAME}-two"
     second_zk_name = f"{ZK_NAME}-two"
 
@@ -49,23 +110,18 @@ async def test_second_cluster(ops_test: OpsTest, kafka_charm, app_charm):
         ops_test.model.deploy(
             ZK_NAME, channel="edge", application_name=second_zk_name, num_units=1, series="jammy"
         ),
-        ops_test.model.deploy(app_charm, application_name=DUMMY_NAME, num_units=1, series="jammy"),
     )
 
     await ops_test.model.wait_for_idle(
-        apps=[second_kafka_name, second_zk_name, DUMMY_NAME],
+        apps=[second_kafka_name, second_zk_name],
         idle_period=30,
         timeout=3600,
     )
     assert ops_test.model.applications[second_kafka_name].status == "blocked"
 
     await ops_test.model.add_relation(second_kafka_name, second_zk_name)
-
-    # Relate "app" to the *first* cluster
-    await ops_test.model.add_relation(APP_NAME, f"{DUMMY_NAME}:{REL_NAME_ADMIN}")
-
     await ops_test.model.wait_for_idle(
-        apps=[second_kafka_name, second_zk_name, DUMMY_NAME, APP_NAME],
+        apps=[second_kafka_name, second_zk_name, APP_NAME],
         idle_period=30,
     )
