+++ tox.ini
@@ -26,7 +26,7 @@ set_env =
     scaling: TEST_FILE=test_scaling.py
     password-rotation: TEST_FILE=test_password_rotation.py
     tls: TEST_FILE=test_tls.py
-    replication: TEST_FILE=test_replication.py
+    ha: TEST_FILE=test_ha.py
 
 pass_env =
     PYTHONPATH
@@ -95,7 +95,7 @@ commands =
     poetry install --with integration
     poetry run pytest -vv --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/{env:TEST_FILE}
 
-[testenv:integration-ha-{replication}]
+[testenv:integration-{ha}]
 description = Run integration tests for high availability
 pass_env =
     {[testenv]pass_env}
